refactor(core): #HttpRequestMetadataStorage
